// src/app/core/enums/entity.enum.ts
export enum EntityEnum {
    AUTHENTICATED_USER = 'AUTHENTICATED_USER',
    PNC = 'PNC',
    CAREER_OBJECTIVE = 'CAREER_OBJECTIVE',
    WAYPOINT = 'WAYPOINT',
    ROTATION = 'ROTATION',
    LEG = 'LEG',
    SUMMARY_SHEET = 'SUMMARY_SHEET',
    CREW_MEMBER = 'CREW_MEMBER',
    PNC_PHOTO = 'PNC_PHOTO',
    EOBSERVATION = 'EOBSERVATION',
    STATUTORY_CERTIFICATE = 'STATUTORY_CERTIFICATE',
    PROFESSIONAL_LEVEL = 'PROFESSIONAL_LEVEL',
    CONGRATULATION_LETTER = 'CONGRATULATION_LETTER',
    PROFESSIONAL_ASSESSMENT = 'PROFESSIONAL_ASSESSMENT'
}